refactor: extract shared observability, services, and models from telegram_bot/ into src/
Fix the src/ to telegram_bot/ layer violation by moving shared code
(observability, BGE-M3 client, VoyageService, apartment models) into
the src/ domain layer. telegram_bot/ files now re-export from src/ for
backward compatibility.

- Create src/observability.py, src/observability_bootstrap.py,
  src/observability_payloads.py with full observability primitives
- Create src/services/ package with bge_m3_client.py, voyage.py,
  _retry.py
- Create src/models/apartment.py with apartment data models
- Update all src/ imports to use new src.* locations
- Make telegram_bot/ modules thin re-export wrappers
- Update test patches to target src.observability and
  src.services.bge_m3_client

Closes #1266
