+++ .agents/tasks/task-fix-layer-violation-1266/context.json
@@ -1,0 +1,38 @@
+{
+  "project_type": "Python monorepo: RAG chatbot with Telegram bot, API, ingestion pipeline",
+  "language": "Python 3.12",
+  "build_system": "uv (pyproject.toml, uv.lock). No custom build step; uses uv sync for deps.",
+  "test_framework": "pytest with pytest-asyncio, pytest-xdist (parallel). Tests in tests/unit/, tests/integration/, tests/contract/. Fixtures in conftest.py files.",
+  "build_command": "uv sync && uv run ruff check src/ telegram_bot/",
+  "test_command": "PYTHONDONTWRITEBYTECODE=1 uv run --python 3.12 pytest tests/unit/ --ignore=tests/unit/test_document_parser.py --ignore=tests/unit/test_evaluator.py --ignore=tests/unit/evaluation/test_ragas_evaluation.py --ignore=tests/unit/voice/test_sip_setup.py --ignore=tests/unit/voice/test_voice_agent.py --ignore=tests/unit/ingestion/test_cocoindex_init.py --ignore=tests/unit/ingestion/test_qdrant_hybrid_target.py --ignore=tests/unit/ingestion/test_qdrant_hybrid_target_helpers.py --ignore=tests/unit/ingestion/test_qdrant_hybrid_target_state_paths.py --ignore=tests/unit/ingestion/test_target_sync_execution.py --ignore=tests/unit/ingestion/test_unified_cli.py --ignore=tests/unit/ingestion/test_unified_flow.py --ignore=tests/unit/ingestion/test_unified_flow_wiring.py -n auto --dist=worksteal -q --timeout=30 -m 'not legacy_api and not requires_extras and not slow'",
+  "verification_instructions": "1. Run `uv run ruff check src/ telegram_bot/` (lint). 2. Run `make test-unit` (core unit tests). 3. Verify no `from telegram_bot` imports remain in src/ files (except src/api/main.py for graph/cache which is out of scope). 4. Verify telegram_bot/ re-exports still work for existing consumers.",
+  "snapshot_or_generated_files": null,
+  "setup_instructions": "1. cd /projects/sandbox/rag. 2. uv sync (installs all deps).",
+  "environment_constraints": "OPEN_INTERNET network mode. Python 3.12 recommended. No Docker daemon needed for unit tests.",
+  "contribution_requirements": "Create branch from dev. Run `make check` and `make test-unit` before claiming completion. Test-first workflow per AGENTS.md.",
+  "key_patterns": "1. Lazy imports used extensively (telegram_bot/services/__init__.py pattern). 2. Observability uses langfuse SDK with graceful degradation (no-op when unavailable). 3. _retry.py provides shared retry decorators. 4. src/ is the domain layer, telegram_bot/ is the transport layer. 5. Tests patch modules at their import location (e.g. patch 'src.ingestion.unified.observability.get_client').",
+  "relevant_files": [
+    "telegram_bot/observability.py",
+    "telegram_bot/observability_bootstrap.py",
+    "telegram_bot/observability_payloads.py",
+    "telegram_bot/services/bge_m3_client.py",
+    "telegram_bot/services/_retry.py",
+    "telegram_bot/services/voyage.py",
+    "telegram_bot/services/apartment_models.py",
+    "telegram_bot/services/__init__.py",
+    "src/ingestion/unified/observability.py",
+    "src/voice/observability.py",
+    "src/retrieval/reranker.py",
+    "src/ingestion/unified/qdrant_writer.py",
+    "src/ingestion/unified/colbert_backfill.py",
+    "src/ingestion/apartments/flow.py",
+    "src/ingestion/apartments/source.py",
+    "src/ingestion/apartments/runner.py",
+    "src/ingestion/cocoindex_flow.py",
+    "src/api/main.py",
+    "src/security/pii_redaction.py",
+    "tests/unit/test_observability.py",
+    "tests/unit/ingestion/test_observability.py"
+  ],
+  "directory_structure": "src/ - domain layer (api/, config/, core/, contextualization/, evaluation/, ingestion/, models/, retrieval/, security/, utils/, voice/). telegram_bot/ - transport/bot layer (agents/, graph/, handlers/, integrations/, services/, middlewares/, pipelines/). tests/ - test suites (unit/, integration/, contract/, smoke/, load/, e2e/, benchmark/). services/ - standalone microservices (bge-m3-api/, docling/, user-base/)."
+}
