+++ tests/unit/test_main.py
@@ -78,7 +78,7 @@ class TestMainFunction:
                     "telegram_bot.logging_config": mock_logging_config_mod,
                 },
             ),
-            patch("telegram_bot.observability._is_endpoint_reachable", return_value=True),
+            patch("src.observability._is_endpoint_reachable", return_value=True),
         ):
             from telegram_bot import main as main_module
 
@@ -118,7 +118,7 @@ class TestMainFunction:
                     "telegram_bot.logging_config": mock_logging_config_mod,
                 },
             ),
-            patch("telegram_bot.observability._is_endpoint_reachable", return_value=True),
+            patch("src.observability._is_endpoint_reachable", return_value=True),
         ):
             from telegram_bot import main as main_module
 
@@ -158,7 +158,7 @@ class TestMainFunction:
                     "telegram_bot.logging_config": mock_logging_config_mod,
                 },
             ),
-            patch("telegram_bot.observability._is_endpoint_reachable", return_value=True),
+            patch("src.observability._is_endpoint_reachable", return_value=True),
         ):
             from telegram_bot import main as main_module
 
@@ -200,7 +200,7 @@ class TestMainFunction:
                     "telegram_bot.logging_config": mock_logging_config_mod,
                 },
             ),
-            patch("telegram_bot.observability._is_endpoint_reachable", return_value=True),
+            patch("src.observability._is_endpoint_reachable", return_value=True),
         ):
             from telegram_bot import main as main_module
 
@@ -248,7 +248,7 @@ class TestMainFunction:
                     "telegram_bot.logging_config": mock_logging_config_mod,
                 },
             ),
-            patch("telegram_bot.observability._is_endpoint_reachable", return_value=True),
+            patch("src.observability._is_endpoint_reachable", return_value=True),
         ):
             from telegram_bot import main as main_module
